pyproject.toml
[tool.poetry]
name = "metrics"
version = "0.1.0"
description = "Collect metrics from Arduino via serial port"
authors = ["Gianluca <gianluca@gurutech.it>"]
license = "GPL-3.0-or-later"
readme = "README.md"
packages = [{include = "metrics"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
pyserial = "^3.5"
pynput = "^1.7.6"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.11.0"
flake8 = "^6.1.0"

[tool.black]
line-length = 200

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
metrics = "metrics.arduino:main"
